simplify makefile hacks for os x

+++ Makefile
@@ -1,23 +1,17 @@
-# Spock GNU Makefile
+# Spock - GNU Makefile
 
-UNAME := $(shell uname)
-
-# ugly OS X hack needed to build bleve with an updated version of libicu
-ifeq ($(UNAME), Darwin)
-ifeq ($(shell brew list icu4c > /dev/null 2>&1; echo $$?), 1)
-$(error You must install icu4c with homebrew)
-endif
-LIBICU_PATH := $(shell dirname `brew list icu4c | fgrep readme.html`)
-export CGO_LDFLAGS := -L$(LIBICU_PATH)/lib
-export CGO_CFLAGS := -I$(LIBICU_PATH)/include
+# Use libicu installation from homebrew
+ifeq ($(shell uname), Darwin)
+export CGO_LDFLAGS := -L/usr/local/opt/icu4c/lib 
+export CGO_CFLAGS := -I/usr/local/opt/icu4c/include
 endif
 
 GO_RICE := $(GOPATH)/bin/rice
-GO_FILES := $(shell find . -name '*.go')
+GO_FILES := *.go cmd/spock/spock.go
 GO_PACKAGE := github.com/piger/spock
 BUILD_TAGS := "libstemmer icu leveldb"
 
-all: install
+all: spock
 
 install: $(GO_FILES) $(GO_RICE)
 	go install -tags $(BUILD_TAGS) $(GO_PACKAGE)/cmd/spock
@@ -27,7 +21,7 @@ spock: $(GO_FILES) $(GO_RICE)
 	go build -tags $(BUILD_TAGS) cmd/spock/spock.go
 
 clean:
-	rm spock
+	test -f spock && rm spock
 
 # Add "-w" to ldflags to skip debug informations.
 release: clean $(GO_RICE)
